+++ internal/command/callbacks.go
@@ -4,19 +4,17 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/chrxn1c/pokemon-repl/internal/user_context"
+	"github.com/chrxn1c/pokemon-repl/internal/entity"
 	"github.com/chrxn1c/pokemon-repl/pkg/api"
 )
 
-type Callback func(ctx *user_context.UserContext) (output string, err error)
-
-func ExitCallback(ctx *user_context.UserContext) (string, error) {
+func ExitCallback(ctx *entity.UserContext) (string, error) {
 	fmt.Println("\nIt's been a pleasure to have you onboard! Thanks for using this application")
 	os.Exit(0)
 	return "", nil
 }
 
-func MapCallback(ctx *user_context.UserContext) (string, error) {
+func MapCallback(ctx *entity.UserContext) (string, error) {
 	ctx.APIoffset += 20
 	endpoint := fmt.Sprintf("location?limit=20&offset=%d", ctx.APIoffset)
 
@@ -32,7 +30,7 @@ func MapCallback(ctx *user_context.UserContext) (string, error) {
 	return toUserResponse, nil
 }
 
-func MapbCallback(ctx *user_context.UserContext) (string, error) {
+func MapbCallback(ctx *entity.UserContext) (string, error) {
 	ctx.APIoffset -= 20
 	endpoint := fmt.Sprintf("location?limit=20&offset=%d", ctx.APIoffset)
