+++ internal/entity/command.go
@@ -1,4 +1,6 @@
-package command
+package entity
+
+type Callback func(ctx *UserContext) (output string, err error)
 
 type Command struct {
 	Name        string `json:"name"`
