+++ internal/entity/userContext.go
@@ -1,12 +1,12 @@
-package user_context
+package entity
 
 import (
-	"github.com/chrxn1c/pokemon-repl/internal/user_context/pokemon"
 	"github.com/chrxn1c/pokemon-repl/pkg/cache"
+	"github.com/chrxn1c/pokemon-repl/pkg/entity"
 )
 
 type UserContext struct {
-	CaughtPokemons []pokemon.Pokemon
+	CaughtPokemons []entity.Pokemon
 	APIoffset      int64
 	Cache          *cache.Cache
 }
