+++ pkg/entity/pokemon.go
@@ -1,4 +1,4 @@
-package pokemon
+package entity
 
 type Pokemon struct {
 	name         string
